Refactor vehicles component
